Get and preview current location (#3)
This PR now allows the user to get their current location. It is converted to an address via the Google Maps Geocoding API.
Notice : Everyone will need their own api-key (or someone else needs to set it up globally).
The address is also saved & displayed in the list items & in the details page.